Rename packages/utils/src/core.ts to src/deps.ts
